requirements.txt
python-docx==0.8.11
PySimpleGUI==4.60.4